Move view state to a separate class
